add types to UIText
