+++ src/components/UIText/index.tsx
@@ -2,16 +2,9 @@ import * as React from 'react';
 import {oneOf, bool, string} from 'prop-types';
 import CoreText from 'wix-ui-core/Text';
 import {ThemedComponent} from 'wix-ui-theme';
-import {theme} from './theme';
+import {theme, Appearance} from './theme';
 
-export type Appearance =
-  'T1' | 'T1.1' | 'T1.2' | 'T1.3' | 'T1.4' | 'T1.5' | 'T1.6' | 'T1.7' |
-  'T2' | 'T2.1' | 'T2.2' | 'T2.3' | 'T2.4' |
-  'T3' | 'T3.1' | 'T3.2' | 'T3.3' | 'T3.4' | 'T3.5' | 'T3.6' | 'T3.7' |
-  'T4' | 'T4.1' | 'T4.2' | 'T4.3' | 'T4.4' | 'T4.5' | 'T4.6' | 'T4.7' |
-  'T5' | 'T5.1';
-
-  interface Props { // TODO: import CoreTextProps adn extend it when wix-ui-core will expose it
+interface Props { // TODO: import CoreTextProps adn extend it when wix-ui-core will expose it
   /** typography of the text */
   appearance?: Appearance;
 
@@ -28,7 +21,7 @@ export type Appearance =
 }
 
 export default class UIText extends React.PureComponent<Props> {
-  propTypes = {
+  static propTypes = {
     ...CoreText.propTypes,
 
     /** typography of the text */
